manticore_cli/requirements.txt
rich>=13.0.0
mysql-connector-python>=8.0.0
